% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_ggplot2.R
\name{EZ_colour_point}
\alias{EZ_colour_point}
\title{Automatique manual scale}
\usage{
EZ_colour_point(primary = "bleu", other = "gris", direction = 1, ...)
}
\arguments{
\item{primary}{first color}

\item{other}{second color}

\item{direction}{direction like ggplot2}

\item{...}{other}
}
\value{
ggplot argument to add to a ggplot object
}
\description{
Automatique manual scale
}
