% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Theme_ggplot2.R
\name{EZ_ggtheme}
\alias{EZ_ggtheme}
\title{ggplot theme}
\usage{
EZ_ggtheme(
  LEGEND = TRUE,
  supress = c("none", "x", "y", "both"),
  rotate = c("none", "x", "y", "both"),
  angle = 90,
  ...
)
}
\arguments{
\item{LEGEND}{permet de spécifier si on veut la présence ou non de la légende.}

\item{supress}{permet de spécifier si on veut supprimer la mise en forme pour un des axes x,y ou les deux.}

\item{...}{other}
}
\value{
a ggplot object
}
\description{
ggplot theme
}
