% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels_ggplot2.R
\name{EZ_label}
\alias{EZ_label}
\title{label for ggplot}
\usage{
EZ_label(
  plot,
  h1_title = NULL,
  y_title = NULL,
  x_title = NULL,
  leg_title = NULL,
  leg_couleur = NULL,
  leg_forme = NULL,
  leg_taille = NULL,
  leg_remplissage = NULL,
  source = NULL,
  champs = NULL,
  subtitle = NULL,
  ...
)
}
\arguments{
\item{plot}{a ggplot object}

\item{h1_title}{main title}

\item{y_title}{titre ordonnée}

\item{x_title}{titre abscisse}

\item{leg_title}{est un moyen facile d'associer un titre sans avoir à spécifier pour quel type de donnée il s'applique}

\item{leg_couleur}{est un moyen d'associer un titre pour les couleurs (color/colour)}

\item{leg_forme}{est un moyen d'associer un titre pour les formes (shape)}

\item{leg_remplissage}{est un moyen d'associer un titre pour les contenants (fill)}

\item{source}{est un moyen d'associer un texte en bas du graphique}

\item{champs}{est un moyen d'associer un texte en bas du graphique}

\item{subtitle}{sous-titre}

\item{...}{other param}
}
\value{
a ggplot object
}
\description{
label for ggplot
}
