% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_ggplot2.R
\name{palette_couleur}
\alias{palette_couleur}
\title{Automatique manual scale}
\usage{
palette_couleur(primary = "bleu", other = "gris", direction = 1)
}
\arguments{
\item{primary}{first color}

\item{other}{second color}

\item{direction}{direction like ggplot2}
}
\value{
character vector
}
\description{
Automatique manual scale
}
